Convert all float fields to scaled integers and currency to DECIMAL
- Updated `init.sql` to change column types.
- Updated SQLAlchemy models (`Planet`, `Fleet`, `Item`, `Contract`) to use `Integer` and `DECIMAL`.
- Updated Pydantic schemas to match model changes.
- Updated `game_data.py` and `galaxy.py` to use scaled integer values (x1000).
- Currency fields (`reward_amount`, `collateral_amount`) set to `DECIMAL(26, 0)`.
- Other physical fields (`x`, `y`, `temperature`, `gravity`, `volume`, `cargo_capacity`) set to `Integer`.

+++ backend/app/models/contract.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey, DateTime, BigInteger
+from sqlalchemy import Column, Integer, String, DECIMAL, ForeignKey, DateTime, BigInteger
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.models.base import Base
@@ -17,8 +17,8 @@ class Contract(Base):
     quantity = Column(Integer, nullable=False)
 
     currency_type = Column(String(10), default="CRED", nullable=False)
-    reward_amount = Column(Float, nullable=False)
-    collateral_amount = Column(Float, nullable=False)
+    reward_amount = Column(DECIMAL(26, 0), nullable=False)
+    collateral_amount = Column(DECIMAL(26, 0), nullable=False)
     duration_seconds = Column(Integer, nullable=False)
 
     created_at = Column(DateTime(timezone=True), server_default=func.now())
